requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

